requirements.txt
fastapi
aiohttp
pillow
uvicorn
pandas
sqlalchemy
databases
asyncpg
